Move test setup logic and test utils into separate files
